perf: remove plugins.rs
